+++ cli/src/pcluster/templates/queues_stack.py
@@ -11,8 +11,10 @@ from pcluster.config.cluster_config import SlurmClusterConfig, SlurmComputeResou
 from pcluster.config.common import DefaultUserHomeType, SharedStorageType
 from pcluster.constants import (
     DEFAULT_EPHEMERAL_DIR,
+    INSTANCE_TYPES_WITH_FIRST_INTERFACE_ENA,
     NODE_BOOTSTRAP_TIMEOUT,
     OS_MAPPING,
+    P6_B300,
     P6E_GB200,
     PCLUSTER_COMPUTE_RESOURCE_NAME_TAG,
     PCLUSTER_QUEUE_NAME_TAG,
@@ -368,9 +370,12 @@ def add_network_interfaces(
     queue_lt_security_groups,
 ):
     """Generate launch template network interfaces list."""
-    is_gb200 = compute_resource.instance_types[0].split(".")[0] == P6E_GB200
+    instance_family = compute_resource.instance_types[0].split(".")[0]
+    is_gb200 = instance_family == P6E_GB200
+    is_b300 = instance_family == P6_B300
     efa_enabled = compute_resource.efa and compute_resource.efa.enabled
-    interface_type = "efa" if efa_enabled and not is_gb200 else None
+    # gb200 and b300 instances require the first interface to be ENA even if EFA is enabled
+    interface_type = "efa" if efa_enabled and instance_family not in INSTANCE_TYPES_WITH_FIRST_INTERFACE_ENA else None
 
     compute_lt_nw_interfaces = [
         ec2.CfnLaunchTemplate.NetworkInterfaceProperty(
@@ -390,10 +395,18 @@ def add_network_interfaces(
         if is_gb200 and not efa_enabled and not even:
             continue
 
-        interface_type = "efa" if efa_enabled else None
-        # if efa is enabled with a gb200 instance, even indexes are configured as efa and the odd as efa-only
-        if is_gb200 and efa_enabled:
-            interface_type = "efa" if even else "efa-only"
+        if efa_enabled:
+            if is_b300:
+                # if efa is enabled with a b300 instance, all network cards, except for the primary,
+                # are configured as efa-only
+                interface_type = "efa-only"
+            elif is_gb200:
+                # if efa is enabled with a gb200 instance, even indexes are configured as efa and the odd as efa-only
+                interface_type = "efa" if even else "efa-only"
+            else:
+                interface_type = "efa"
+        else:
+            interface_type = None
 
         compute_lt_nw_interfaces.append(
             ec2.CfnLaunchTemplate.NetworkInterfaceProperty(